Build cc1 with MSVC, and let Visual Studio compile C with it
Two things, and the second is the point: a Visual Studio project whose C
files are compiled by this compiler rather than by cl. msvc\cc1-as-cl.bat
is named in CLToolExe and MSBuild runs it wherever it would have run
cl.exe, exactly as tools\cc1-as-clang lets Xcode do the same on the Mac.

Nothing in src\ is restructured for MSVC. The compiler's only dependency
outside C++17 was getpid, and msvc\compat\unistd.h answers that include
with the directory placed first on the include path - a shim in the
project that wants it rather than a platform #ifdef in a file that is
about compiling C.

**The port found two real bugs, and they are fixed rather than silenced.**
Both are LLP64: unsigned long is 32 bits under MSVC and 64 everywhere else
this is built, so neither gcc nor clang could ever have reported them.

  - x87Parts held an 80-bit significand in an unsigned long, so a
    Windows-hosted cc1 would have written truncated long double constants
    when cross-compiling to x86_64-linux.
  - Parser.cpp guarded LONG_MIN / -1 with (1UL << 63), which is undefined
    when unsigned long is 32 bits.

Auditing for the pattern rather than the two MSVC could prove found more,
including std::memcpy(&bits, &d, 8) into a four-byte object - an eight-byte
write into a stack variable half that size. Every value that is a 64-bit
bit pattern is an unsigned long long now. On Linux and macOS the types were
already 64 bits, so nothing moved: 421 Linux, 18 Windows-GNU, 16 arm64, and
all 412 cases still agree with clang exactly on arm64.

The toolset is deliberately not pinned. Writing v143 names the 2022 build
tools exactly and a machine with a later Visual Studio refuses the project
outright with MSB8020, which is what happened on the machine this was first
built on. $(DefaultPlatformToolset) is v143 under 2022 and whatever is
installed elsewhere, so the project stays 2022-compatible without being
2022-only.

The shim reads its flags case-sensitively, and that is not fussiness: cl
has /D for a macro and /diagnostics:column for a message format, and a
case-blind pattern read the second as defining 'iagnostics:column'. It also
expands MSBuild's response file, without which it sees one argument
beginning with @ and no source at all.

readme.txt is the procedure, and every step in it was run from a clean
directory before being written - including the failure cases: windows.h
refused with the path it looked in, and the substitution removed again to
show cl building the same program to the same output.

+++ src/Parser.cpp
@@ -876,7 +876,7 @@ ExprPtr Parser::primary(Program *program) {
     if (peek().kind == TokenKind::Num) {
         const Token &t = peek();
         const Type *ty;
-        unsigned long u = static_cast<unsigned long>(t.value);
+        unsigned long long u = static_cast<unsigned long long>(t.value);
         // LL before L: 'long long' is a distinct type and on a target where
         // long is 32 bits it is the only one of the two wide enough. Typing
         // 42LL as long would narrow it on x86_64-windows and nowhere else.
@@ -1352,7 +1352,7 @@ void Parser::flattenScalar(const Type *type, Init &in, int base,
             // significand, then the sign and exponent two bytes further on.
             // The six bytes of padding after them are left for the emitter to
             // zero, which is what it does with any gap.
-            unsigned long sig = 0;
+            unsigned long long sig = 0;
             unsigned int hi = 0;
             x87Parts(d, &sig, &hi);
             out.push_back(GlobalPiece{ base, 8, static_cast<long>(sig),
@@ -1368,7 +1368,7 @@ void Parser::flattenScalar(const Type *type, Init &in, int base,
             bits = static_cast<long>(u);
         } else {
             double dd = static_cast<double>(d);
-            unsigned long u;
+            unsigned long long u;
             std::memcpy(&u, &dd, sizeof u);
             bits = static_cast<long>(u);
         }
@@ -2331,7 +2331,7 @@ bool Parser::fold(const Expr &e, long *out, std::size_t pos) const {
         long v;
         if (!fold(u->operand(), &v, pos)) return false;
         switch (u->op()) {
-        case '-': *out = static_cast<long>(0UL - static_cast<unsigned long>(v)); return true;
+        case '-': *out = static_cast<long>(0ULL - static_cast<unsigned long long>(v)); return true;
         case '+': *out = v; return true;
         case '!': *out = !v; return true;
         case '~': *out = ~v; return true;
@@ -2351,8 +2351,8 @@ bool Parser::fold(const Expr &e, long *out, std::size_t pos) const {
 
         const Type *t = b->lhs().type();
         bool uns = t->isInteger() && !t->isSigned(target_);
-        unsigned long ul = static_cast<unsigned long>(l);
-        unsigned long ur = static_cast<unsigned long>(r);
+        unsigned long long ul = static_cast<unsigned long long>(l);
+        unsigned long long ur = static_cast<unsigned long long>(r);
 
         switch (b->op()) {
         case BinOp::Add: *out = static_cast<long>(ul + ur); return true;
@@ -2362,7 +2362,7 @@ bool Parser::fold(const Expr &e, long *out, std::size_t pos) const {
         case BinOp::Mod:
             if (r == 0)
                 src_.fail(pos, "division by zero in a constant expression");
-            if (!uns && ul == (1UL << 63) && r == -1) {
+            if (!uns && ul == (1ULL << 63) && r == -1) {
                 *out = (b->op() == BinOp::Div) ? l : 0;
                 return true;
             }
@@ -2399,9 +2399,9 @@ bool Parser::fold(const Expr &e, long *out, std::size_t pos) const {
 long Parser::narrowTo(long v, const Type *t) const {
     int bits = t->size(target_) * 8;
     if (bits >= 64) return v;
-    unsigned long mask = (1UL << bits) - 1;
-    unsigned long kept = static_cast<unsigned long>(v) & mask;
-    if (t->isSigned(target_) && (kept & (1UL << (bits - 1)))) kept |= ~mask;
+    unsigned long long mask = (1ULL << bits) - 1;
+    unsigned long long kept = static_cast<unsigned long long>(v) & mask;
+    if (t->isSigned(target_) && (kept & (1ULL << (bits - 1)))) kept |= ~mask;
     return static_cast<long>(kept);
 }
